Move Graphql Schema
- Belongs in routes folder and should have been moved a while ago
