+++ server/src/routes/graphql.js
@@ -11,19 +11,19 @@ const scanAreas = fs.existsSync('server/src/configs/areas.json')
   ? require('../configs/areas.json')
   : { features: [] }
 
-const DeviceType = require('./device')
-const GeocoderType = require('./geocoder')
-const GymType = require('./gym')
-const NestType = require('./nest')
-const PokestopType = require('./pokestop')
-const PokemonType = require('./pokemon')
-const PoracleType = require('./poracle')
-const PortalType = require('./portal')
-const S2cellType = require('./s2cell')
-const ScanAreaType = require('./scanArea')
-const SearchType = require('./search')
-const SpawnpointType = require('./spawnpoint')
-const WeatherType = require('./weather')
+const DeviceType = require('../schema/device')
+const GeocoderType = require('../schema/geocoder')
+const GymType = require('../schema/gym')
+const NestType = require('../schema/nest')
+const PokestopType = require('../schema/pokestop')
+const PokemonType = require('../schema/pokemon')
+const PoracleType = require('../schema/poracle')
+const PortalType = require('../schema/portal')
+const S2cellType = require('../schema/s2cell')
+const ScanAreaType = require('../schema/scanArea')
+const SearchType = require('../schema/search')
+const SpawnpointType = require('../schema/spawnpoint')
+const WeatherType = require('../schema/weather')
 const Utility = require('../services/Utility')
 const Fetch = require('../services/Fetch')
 const config = require('../services/config')
